+++ src/components/Product.jsx
@@ -1,11 +1,15 @@
-export const Product = ({ product }) => {
-	return (
-		<div className="product" key={product.id}>
-			<img src={product.img} alt={product.name} />
-			<div className="product__text">
-				<p className="product__title">{product.name}</p>
-				<p className="product__price">{product.price}</p>
-			</div>
-		</div>
-	);
-};
+import snowboard from "../images/snowboard_light.png";
+
+export const Product = () => {
+    return (
+        <div className="product">
+            <img src={snowboard} />
+            <div className="product__content">
+
+
+
+                <p>title</p>
+            </div>
+        </div>
+    )
+}
