+++ src/components/ProductCard.jsx
@@ -1,40 +1,11 @@
-import React from "react";
-import PropTypes from "prop-types";
-
-const ProductCard = ({ product }) => {
+export const ProductCard = ({ product }) => {
 	return (
-		<div className="product-card">
-			<div className="product-image">
-				<img src={product.imageUrl} alt={product.name} />
-			</div>
-			<div className="product-details">
-				<h3 className="product-name">{product.name}</h3>
-				<p className="product-brand">
-					Brand: {product.brand}
-				</p>
-				<p className="product-price">
-					Price: ${product.price.toFixed(2)}
-				</p>
-				<button
-					className="add-to-cart-btn"
-					disabled={!product.stock}
-				>
-					Add to Cart
-				</button>
+		<div className="product" key={product.id}>
+			<img src={product.img} alt={product.name} />
+			<div className="product__text">
+				<p className="product__title">{product.name}</p>
+				<p className="product__price">{product.price}</p>
 			</div>
 		</div>
 	);
 };
-
-ProductCard.propTypes = {
-	product: PropTypes.shape({
-		id: PropTypes.string.isRequired,
-		name: PropTypes.string.isRequired,
-		brand: PropTypes.string.isRequired,
-		imageUrl: PropTypes.string.isRequired,
-		price: PropTypes.number.isRequired,
-		stock: PropTypes.number.isRequired,
-	}).isRequired,
-};
-
-export default ProductCard;
