+++ src/scss/layout/_product.scss
@@ -1,0 +1,4 @@
+.product {
+    display: flex;
+    flex-direction: column;
+}
