+++ src/scss/main.scss
@@ -5,7 +5,7 @@
 @import "utils/accordion";
 @import "utils/button";
 @import "utils/form";
-@import "utils/product";
+@import "utils/product-card";
 
 @import "layout/wrapper";
 @import "layout/account";
